Remove v1 tracer and move test to core
