+++ langchain-core/.gitignore
@@ -85,9 +85,6 @@ tracers/log_stream.d.ts
 tracers/run_collector.cjs
 tracers/run_collector.js
 tracers/run_collector.d.ts
-tracers/tracer_langchain_v1.cjs
-tracers/tracer_langchain_v1.js
-tracers/tracer_langchain_v1.d.ts
 tracers/tracer_langchain.cjs
 tracers/tracer_langchain.js
 tracers/tracer_langchain.d.ts
