+++ langchain-core/src/load/import_map.ts
@@ -28,7 +28,6 @@ export * as tracers__console from "../tracers/console.js";
 export * as tracers__initialize from "../tracers/initialize.js";
 export * as tracers__log_stream from "../tracers/log_stream.js";
 export * as tracers__run_collector from "../tracers/run_collector.js";
-export * as tracers__tracer_langchain_v1 from "../tracers/tracer_langchain_v1.js";
 export * as tracers__tracer_langchain from "../tracers/tracer_langchain.js";
 export * as utils__async_caller from "../utils/async_caller.js";
 export * as utils__chunk_array from "../utils/chunk_array.js";
