+++ langchain-core/src/tracers/tests/langchain_tracer.int.test.ts
@@ -1,0 +1,43 @@
+/* eslint-disable no-process-env */
+import * as uuid from "uuid";
+import { test } from "@jest/globals";
+
+import { LangChainTracer } from "../tracer_langchain.js";
+import { Serialized } from "../../load/serializable.js";
+import { HumanMessage } from "../../messages/index.js";
+
+const serialized: Serialized = {
+  lc: 1,
+  type: "constructor",
+  id: ["test"],
+  kwargs: {},
+};
+
+test("LangChain V2 tracer does not throw errors for its methods", async () => {
+  const tracer = new LangChainTracer({
+    projectName: `JS Int Test - ${uuid.v4()}`,
+  });
+  const chainRunId = uuid.v4();
+  const toolRunId = uuid.v4();
+  const llmRunId = uuid.v4();
+  const chatRunId = uuid.v4();
+  await tracer.handleChainStart(serialized, { foo: "bar" }, chainRunId);
+  await tracer.handleToolStart(serialized, "test", toolRunId, chainRunId);
+  await tracer.handleLLMStart(serialized, ["test"], llmRunId, toolRunId);
+  await tracer.handleLLMEnd({ generations: [[]] }, llmRunId);
+  await tracer.handleChatModelStart(
+    serialized,
+    [[new HumanMessage("I'm a human.")]],
+    chatRunId
+  );
+  await tracer.handleLLMEnd({ generations: [[]] }, chatRunId);
+  await tracer.handleToolEnd("output", toolRunId);
+  const llmRunId2 = uuid.v4();
+  await tracer.handleLLMStart(serialized, ["test"], llmRunId2, chainRunId);
+  await tracer.handleLLMEnd({ generations: [[]] }, llmRunId2);
+  await tracer.handleChainEnd({ foo: "bar" }, chainRunId);
+
+  const llmRunId3 = uuid.v4();
+  await tracer.handleLLMStart(serialized, ["test"], llmRunId3);
+  await tracer.handleLLMEnd({ generations: [[]] }, llmRunId3);
+});
