+++ langchain/src/callbacks/index.ts
@@ -13,12 +13,7 @@ export { RunCollectorCallbackHandler } from "./handlers/run_collector.js";
 
 export { LangChainTracer } from "./handlers/tracer_langchain.js";
 
-export { LangChainTracerV1 } from "./handlers/tracer_langchain_v1.js";
-
-export {
-  getTracingCallbackHandler,
-  getTracingV2CallbackHandler,
-} from "./handlers/initialize.js";
+export { getTracingV2CallbackHandler } from "./handlers/initialize.js";
 
 export {
   CallbackManager,
